+++ spatialos-sdk/src/worker/mod.rs
@@ -177,25 +177,6 @@ pub enum ConnectionStatusCode {
     ServerShutdown,
 }
 
-impl From<u8> for ConnectionStatusCode {
-    fn from(value: u8) -> Self {
-        match u32::from(value) {
-            1 => ConnectionStatusCode::Success,
-            2 => ConnectionStatusCode::InternalError,
-            3 => ConnectionStatusCode::InvalidArgument,
-            4 => ConnectionStatusCode::NetworkError,
-            5 => ConnectionStatusCode::Timeout,
-            6 => ConnectionStatusCode::Cancelled,
-            7 => ConnectionStatusCode::Rejected,
-            8 => ConnectionStatusCode::PlayerIdentityTokenExpired,
-            9 => ConnectionStatusCode::LoginTokenExpired,
-            10 => ConnectionStatusCode::CapacityExceeded,
-            11 => ConnectionStatusCode::RateExceeded,
-            12 => ConnectionStatusCode::ServerShutdown,
-            _ => panic!(format!("Unknown connection status code: {}", value)),
-        }
-    }
-}
 
 pub mod constants {
     pub const LOCATOR_HOSTNAME: &str = "locator.improbable.io";
